Crate iroh_node_util

Utilities for building iroh nodes.

Modules

clicli

Cli commands.

configconfig

Utilities to get default paths for configuration, data, and cache directories.

fs

Utilities for filesystem operations.

logginglogging

Utilities for logging

rpc

RPC client, server and protocol to control iroh endpoints and iroh nodes