Module iroh_node_util::rpc::proto::net

RPC calls to control an iroh endpoint.

Structs

AddAddrRequest

AddrRequest

IdRequest

A request to get information the identity of the node.

NodeWatchRequest

A request to watch for the node status

RelayRequest

RemoteInfoRequest

Get information about a specific remote node.

RemoteInfoResponse

A response to a Request::RemoteInfo request

RemoteInfosIterRequest

List network path information about all the remote nodes known by this node.

RemoteInfosIterResponse

A response to a Request::RemoteInfosIter.

WatchResponse

The response to a watch request

Enums

Request

Response