iroh_node_util::rpc
Module server

Server implementation to handle node and net rpc requests

Traits

AbstractNode

Trait that provides fields used by the rpc handler for the net and node requests.

Functions

handle_rpc_request

Handle rpc requests for the node and net services